# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEvent {
  id: number;
  customerId: number;
  customerName: string;
  productId: number;
  productName: string;
  type: string;
  occurredAt: Date;
}