# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityInputType } from './activityInputType';

export interface ActivityInput {
  customerId: number;
  productId: number;
  type: ActivityInputType;
}