# lib/api-zod/src/generated/types/analyticsDashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChartPoint } from './chartPoint';
import type { Metric } from './metric';

export interface AnalyticsDashboard {
  metrics: Metric[];
  topProducts: ChartPoint[];
  categoryRevenue: ChartPoint[];
  monthlyRevenue: ChartPoint[];
  segments: ChartPoint[];
  orderStatuses: ChartPoint[];
}