# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  customerId: number;
  status: string;
  total: number;
  paymentMethod: string;
  paymentStatus: string;
  placedAt: Date;
  items: OrderItem[];
}