# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputPaymentMethod } from './orderInputPaymentMethod';
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  customerId: number;
  /** @minItems 1 */
  items: OrderItemInput[];
  paymentMethod: OrderInputPaymentMethod;
}